rdma_cm_client: Add a bind option for client side rdma_cm
In certain scenarios it is desirable to be able to bind the client
side of a rdma_cm based  connection. For example, in certain loopback
configurations or when multiple HCAs exist in the machine and more
than one has a path to the server.

Add a -L <client name> option that allows the user to specify which
local client to pass to a rdma_bind_addr() function call. Omit this
option to bind to the first valid address.